refacto: moved ws.route out of folder routes and renamed folder handlers to game
